aurora_pro/requirements.txt
aiolimiter>=1.1.0
aiosqlite>=0.19.0
aiofiles>=23.1.0
fastapi>=0.104.0
fpdf2>=2.7.0
httpx>=0.25.0
mss>=9.0.1
imageio>=2.31.0
newspaper3k>=0.2.8
numpy>=1.24.0
opencv-python>=4.9.0.80
pandas>=2.1.0
Pillow>=10.0.0
plotly>=5.18.0
psutil>=5.9.7
prometheus-client>=0.19.0
python-multipart>=0.0.6
pydantic>=2.5.0
selenium>=4.18.0
SpeechRecognition>=3.10.0
streamlit>=1.35.0
streamlit-ace>=0.1.1
streamlit-chat>=0.1.1
streamlit-elements==0.1.0
tenacity>=8.2.0
trafilatura>=1.6.0
uvicorn[standard]>=0.24.0
websockets>=12.0
webdriver-manager>=4.0.0
requests>=2.31.0
altair>=5.0.0
pyautogui>=0.9.54
PyYAML>=6.0

# Advanced Automation & Stealth
undetected-chromedriver>=3.5.5
selenium-stealth>=1.0.6
2captcha-python>=1.2.1

# Vision & OCR
pytesseract>=0.3.10

# Advanced Networking & Performance
aiohttp>=3.9.0
aiohappyeyeballs>=2.6.0
aiosignal>=1.4.0
frozenlist>=1.7.0
multidict>=6.6.0
propcache>=0.3.0
yarl>=1.20.0
diskcache>=5.6.3
redis>=5.0.1

# System Optimization
py-cpuinfo>=9.0.0
pynput>=1.7.6
